Adding Login, Loading, and About Screen UI's
Built the UI for the Login, Loading, and About screens based off the templates shown in the ArcPlanner UI Design Template document in the project's wiki. These are the visuals only with very little logic behind them at the moment. The backend logic for each screen will be added in with the rest of the app functionality at a later date.