Adding Screen Coding Specification Document
Added the ArcPlanner Screen Coding Specifications document to the branch. The document is a WIP but will remain that way for the forseeable future. The purpose of the document is to start conversation about the app's UI style and is open to discussion and changes. This should the place to come to see how certain aspects of screens are built in Flutter. If you come across an aspect of screen design that isn't touched upon in this document, feel free to make a section for it and start a conversation about it.