Add Home Screen and Side Menu
Added the UI build for the Home screen. At current there is no implementation of a task tile as there are no current test tasks.

Also adds in the Side Menu to the Home and About screens. This allows for navigation to completed screens in the app, currently only the Home, About, and Login screens. The Loading screen is unreachable at current as there are no operations that require signifigant load times.