Update Screens to Prepare for BLoC Patterning
Updates the UI files to use StatelessWidget objects instead of StatefulWidget objects. This allows us to incorporate the BLoC pattern into ArcPlanner.

This commit also fixes minor issues in the screens such as unused imports and separates out some lengthy widget definitions into functions.

In addition, the commit establishes the arc_view_screen.dart which is currently non-functional, as well as establishes the arc_tree class. The arc_tree class is a possible solution to storing data pulled from the DB for use in the client-facing app.